split misc module into auth and non-auth for better modularity
